+++ trunk/mts/src/main/java/com/devoteam/srit/xmlloader/gtp/data/HeaderGTPPrime.java
@@ -92,10 +92,10 @@ public class HeaderGTPPrime extends HeaderAbstract
         if (strType != null)
         {
             EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-            this.type = field.getEnumValue(strType);
+            this.type = (int) field.getEnumValue(strType);
         }
         EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-        this.label = field.getEnumLabelByValue(this.type);
+        this.label = field.getEnumLabelByValue((long) this.type);
         
         String attribute;
         String attrFlag;
@@ -158,7 +158,7 @@ public class HeaderGTPPrime extends HeaderAbstract
 		int offset = 4;
 		
     	EnumerationField field = (EnumerationField) dictionary.getHeaderFieldByName("Message Type");
-    	this.label = field.getEnumLabelByValue(this.type);
+    	this.label = field.getEnumLabelByValue((long) this.type);
     	
     	Array seqnumArray = array.subArray(offset, 2); 	
     	this.sequenceNumber = (new Integer16Array(seqnumArray).getValue());
